+++ CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - (cli) Added tracking of the last pallet path@version query used with the `plt clone` and `plt switch` subcommands, so that those subcommands can be called again with a partial query (i.e. `@version_query` or `pallet_path@` or `@`) to reuse the last provided value(s) for omitted parts of the query.
-- (cli) Added a `plt upgrade` subcommand as a more user-friendly equivalent to `plt switch @`.
-- (cli) Added `plt show-upgrade-query` to show the pallet path@version query which will be used for `plt upgrade` and for `plt clone/switch` subcommands with partial queries.
-- (cli) Added `plt set-upgrade-query` to modify the pallet path@version query which will be used for `plt upgrade` and for `plt clone/switch` subcommands with partial queries.
+- (cli) Added a `plt upgrade` subcommand as a upgrade-specific version of `plt switch` (with additional checks and log messages).
+- (cli) Added a `plt check-upgrade` subcommand to show whether an upgrade is available and, if so, what change to the local pallet would be made by `plt upgrade`.
+- (cli) Added a `plt show-upgrade-query` subcommand to show the pallet path@version query which will be used for `plt upgrade` and for `plt clone/switch` subcommands with partial queries.
+- (cli) Added a `plt set-upgrade-query` subcommand to modify the pallet path@version query which will be used for `plt upgrade` and for `plt clone/switch` subcommands with partial queries.
 
 ### Fixed